MFW when the

+++ src/d_netcmd.c
@@ -917,7 +917,7 @@ boolean EnsurePlayerNameIsGood(char *name, INT32 playernum)
 	// differ on whether they're printable characters or not.
 	for (ix = 0; name[ix] != '\0'; ix++)
 	{
-		UINT8 printable = isprint(name[ix]);
+		int printable = isprint(name[ix]);
 		// Lactozilla: Apparently Unicode isn't printable.
 		// Oh well.
 		if ((UINT8)name[ix] > 0x7F)
